uniformize radius type and default values

+++ clic/src/tier1/mode.cpp
@@ -13,9 +13,9 @@ auto
 mode_func(const Device::Pointer & device,
           const Array::Pointer &  src,
           Array::Pointer          dst,
-          int                     radius_x,
-          int                     radius_y,
-          int                     radius_z,
+          float                     radius_x,
+          float                     radius_y,
+          float                     radius_z,
           std::string             connectivity) -> Array::Pointer
 {
   if (src->dtype() != dType::UINT8)
@@ -43,9 +43,9 @@ auto
 mode_sphere_func(const Device::Pointer & device,
                  const Array::Pointer &  src,
                  Array::Pointer          dst,
-                 int                     radius_x,
-                 int                     radius_y,
-                 int                     radius_z) -> Array::Pointer
+                 float                     radius_x,
+                 float                     radius_y,
+                 float                     radius_z) -> Array::Pointer
 {
   return mode_func(device, src, dst, radius_x, radius_y, radius_z, "sphere");
 }
@@ -54,9 +54,9 @@ auto
 mode_box_func(const Device::Pointer & device,
               const Array::Pointer &  src,
               Array::Pointer          dst,
-              int                     radius_x,
-              int                     radius_y,
-              int                     radius_z) -> Array::Pointer
+              float                     radius_x,
+              float                     radius_y,
+              float                     radius_z) -> Array::Pointer
 {
   return mode_func(device, src, dst, radius_x, radius_y, radius_z, "box");
 }
